the maps are being changed s
when a restart the game they are playing with the maps modified, so i have to clone all the game Maps. 
Solution: override clone() in GameMap as abstract function and do the clone automaticaly when the GameState COnstrutcter is called. When i restart the game just coppy them again to the arrayList